requirements.txt
annotated-types==0.7.0
anyio==4.5.0
backports.zoneinfo==0.2.1
certifi==2024.8.30
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
jiter==0.5.0
openai==1.50.2
psycopg2==2.9.9
pydantic==2.9.2
pydantic-core==2.23.4
pypdf==5.0.0
python-dotenv==1.0.1
sniffio==1.3.1
tqdm==4.66.5
typing-extensions==4.12.2
tzdata==2024.2